Changed input variable

+++ modules/calc_module.h
@@ -1,6 +1,6 @@
 #ifndef CALC_MODULE_H_
 #define CALC_MODULE_H_
 
-float pi_count(int x);
+float pi_count(float x);
 
 #endif
